chore(frontend): reorganize frontend code
