+++ frontend/src/components/Cell.svelte
