+++ frontend/src/lib/CommonClient.ts
@@ -1,4 +1,4 @@
-import { Empty } from './gen/common_pb';
+import { Empty } from "../gen/common_pb";
 
 export function newEmpty(): Empty {
     return new Empty();
